+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This document details the changes between each release.
 ### Added
 * Added a way to set and retrieve 16-bit values using new `Receiver::get16Bit`
   and `Sender::set16Bit` functions.
+* Added the ability to enable and disable the transmitter in the receiver via
+  a new `Receiver::setTXEnabled` function. This is useful when it is known that
+  the receiver will be receive-only and it is not desired to drive the TX line.
 
 ### Changed
 * In the sender, changed how the baud rate is set when switching between the